Move planner notes to separate file
